+++ static/js/statsvar_menu.tsx
